#5383: Rename demos/falcon7b to demos/falcon7b_common
