+++ models/demos/falcon7b_common/README.md
@@ -1,0 +1,10 @@
+# Falcon7B
+
+This folder contains Falcon7b code which is shared by the Grayskull, Wormhole, and T3000 demos, as well as other Falcon7b tests. To run the Falcon7b demos, please navigate to the links below:
+
+
+[Falcon7B Grayskull Demo](https://github.com/tenstorrent/tt-metal/tree/main/models/demos/grayskull/falcon7b)
+
+[Falcon7B Wormhole Demo](https://github.com/tenstorrent/tt-metal/tree/main/models/demos/wormhole/falcon7b)
+
+[Falcon7B T3000 Demo](https://github.com/tenstorrent/tt-metal/tree/main/models/demos/t3000/falcon7b)
