+++ models/demos/falcon7b_common/reference/cpu_demo.py
@@ -9,7 +9,7 @@ from loguru import logger
 
 from transformers import AutoTokenizer
 
-from models.falcon7b.reference.hf_modeling_falcon import FalconForCausalLM
+from models.demos.falcon7b_common.reference.hf_modeling_falcon import FalconForCausalLM
 import time
 
 falcon1b = "tiiuae/falcon-rw-1b"
