+++ models/demos/falcon7b_common/tests/test_falcon_end_to_end.py
@@ -7,16 +7,16 @@ import pytest
 import torch
 import tt_lib
 from loguru import logger
-from models.demos.falcon7b.tests.test_utils import (
+from models.demos.falcon7b_common.tests.test_utils import (
     concat_device_out_layer_present,
     get_rand_falcon_inputs,
     load_hf_model,
 )
-from models.demos.falcon7b.tt.falcon_causallm import TtFalconCausalLM
+from models.demos.falcon7b_common.tt.falcon_causallm import TtFalconCausalLM
 
 # TODO: Remove this?
-from models.demos.falcon7b.tt.falcon_common import PytorchFalconCausalLM
-from models.demos.falcon7b.tt.model_config import get_model_config
+from models.demos.falcon7b_common.tt.falcon_common import PytorchFalconCausalLM
+from models.demos.falcon7b_common.tt.model_config import get_model_config
 from models.utility_functions import (
     disable_compilation_reports,
     disable_persistent_kernel_cache,
