+++ models/demos/falcon7b_common/tests/test_falcon_mlp.py
@@ -5,9 +5,9 @@
 import pytest
 import torch
 from loguru import logger
-from models.demos.falcon7b.tt.falcon_mlp import TtFalconMLPDecode, TtFalconMLPPrefill
-from models.demos.falcon7b.tt.model_config import get_model_config
-from models.demos.falcon7b.tests.test_utils import load_hf_model
+from models.demos.falcon7b_common.tt.falcon_mlp import TtFalconMLPDecode, TtFalconMLPPrefill
+from models.demos.falcon7b_common.tt.model_config import get_model_config
+from models.demos.falcon7b_common.tests.test_utils import load_hf_model
 from models.utility_functions import get_devices_for_t3000, torch2tt_tensor, tt2torch_tensor
 from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_allclose, comp_pcc
