+++ models/demos/falcon7b_common/tests/test_falcon_prefill_decode.py
@@ -6,12 +6,12 @@ import torch
 import pytest
 from loguru import logger
 
-from models.demos.falcon7b.tt.falcon_causallm import TtFalconCausalLM
+from models.demos.falcon7b_common.tt.falcon_causallm import TtFalconCausalLM
 
-from models.demos.falcon7b.tt.model_config import (
+from models.demos.falcon7b_common.tt.model_config import (
     get_model_config,
 )
-from models.demos.falcon7b.tests.test_utils import load_hf_model
+from models.demos.falcon7b_common.tests.test_utils import load_hf_model
 
 from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import (
     comp_allclose,
