+++ models/demos/falcon7b_common/tests/unit_tests/test_falcon_matmuls_and_bmms_with_mixed_precision.py
@@ -7,9 +7,9 @@ from enum import Enum, auto
 from loguru import logger
 
 import ttnn
-from models.demos.falcon7b.tt.falcon_causallm import falcon_lm_head_matmul
-from models.demos.falcon7b.tt.falcon_mlp import falcon_dense_4h_to_h_matmul, falcon_dense_h_to_4h_matmul
-from models.demos.falcon7b.tt.model_utils import get_falcon_default_core_grid
+from models.demos.falcon7b_common.tt.falcon_causallm import falcon_lm_head_matmul
+from models.demos.falcon7b_common.tt.falcon_mlp import falcon_dense_4h_to_h_matmul, falcon_dense_h_to_4h_matmul
+from models.demos.falcon7b_common.tt.model_utils import get_falcon_default_core_grid
 from models.utility_functions import comp_pcc, tt2torch_tensor, torch2tt_tensor, skip_for_wormhole_b0
 import torch
 import math
