+++ models/demos/falcon7b_common/tt/falcon_mlp.py
@@ -4,7 +4,7 @@
 
 import torch
 import ttnn
-from models.demos.falcon7b.tt.model_utils import get_falcon_default_core_grid, get_weights_cached
+from models.demos.falcon7b_common.tt.model_utils import get_falcon_default_core_grid, get_weights_cached
 from torch import nn
 from models.utility_functions import (
     is_grayskull,
